+++ pdf_handler_test.go
@@ -2,15 +2,13 @@
 package lazypdf
 
 import (
-	"bytes"
 	"context"
 	"fmt"
-	"io"
 	"log/slog"
-	"math"
 	"os"
-	"strings"
+	"path/filepath"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/require"
 )
@@ -80,7 +78,7 @@ func TestPdfHandler_TestClosePDF(t *testing.T) {
 		t.Fatalf("ClosePDF: %v", err)
 	}
 }
-func TestPdfHandler_LocationSizeToPdfPoints_InvalidPage(t *testing.T) {
+func TestPdfHandler_ConvertTopLeftToBottomLeft_InvalidPage(t *testing.T) {
 	t.Parallel()
 
 	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
@@ -96,12 +94,10 @@ func TestPdfHandler_LocationSizeToPdfPoints_InvalidPage(t *testing.T) {
 	}
 	defer func() { require.NoError(t, handler.ClosePDF(document)) }()
 
-	//nolint:dogsled // we only care about the error
-	_, _, _, _, err = handler.LocationSizeToPdfPoints(
+	_, _, err = handler.ConvertTopLeftToBottomLeft(
 		context.Background(),
 		document,
 		2,
-		0,
 		0,
 		0,
 		0,
@@ -111,117 +107,53 @@ func TestPdfHandler_LocationSizeToPdfPoints_InvalidPage(t *testing.T) {
 	require.Equal(t, "failed to get page size: failure at the C/MuPDF get_page_size function: invalid page number: 3", err.Error())
 }
 
-func TestPdfHandler_LocationSizeToPdfPoints_InvalidInputPercentages(t *testing.T) {
+func TestPdfHandler_ConvertTopLeftToBottomLeft(t *testing.T) {
 	t.Parallel()
 
 	tests := []struct {
-		name                string
-		path                string
-		x, y, width, height float64
-		expectedError       string
+		name      string
+		path      string
+		x         float64
+		y         float64
+		height    float64
+		expectedX float64
+		expectedY float64
 	}{
 		{
-			name:          "X greater than 1",
-			path:          "testdata/pdf_handler_sample.pdf",
-			x:             1.1,
-			y:             0.5,
-			width:         0.5,
-			height:        0.5,
-			expectedError: "invalid input percentages: x=1.100000, y=0.500000, width=0.500000, height=0.500000",
-		},
-		{
-			name:          "Y less than 0",
-			path:          "testdata/pdf_handler_sample.pdf",
-			x:             0.5,
-			y:             -0.1,
-			width:         0.5,
-			height:        0.5,
-			expectedError: "invalid input percentages: x=0.500000, y=-0.100000, width=0.500000, height=0.500000",
-		},
-		{
-			name:          "Width less than 0",
-			path:          "testdata/pdf_handler_sample.pdf",
-			x:             0.5,
-			y:             0.5,
-			width:         -0.1,
-			height:        0.5,
-			expectedError: "invalid input percentages: x=0.500000, y=0.500000, width=-0.100000, height=0.500000",
-		},
-		{
-			name:          "Height greater than 1",
-			path:          "testdata/pdf_handler_sample.pdf",
-			x:             0.5,
-			y:             0.5,
-			width:         0.5,
-			height:        1.1,
-			expectedError: "invalid input percentages: x=0.500000, y=0.500000, width=0.500000, height=1.100000",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			handler := setupPdfHandler(t)
-			document := openTestPDF(t, tt.path)
-
-			_, _, _, _, err := handler.LocationSizeToPdfPoints(context.Background(), document, 0, tt.x, tt.y, tt.width, tt.height)
-			require.Error(t, err)
-			require.EqualError(t, err, tt.expectedError)
-		})
-	}
-}
-
-func TestPdfHandler_TestLocationSizeToPdfPoints(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name           string
-		path           string
-		X              float64
-		Y              float64
-		Width          float64
-		Height         float64
-		expectedX      float64
-		expectedY      float64
-		expectedWidth  float64
-		expectedHeight float64
-	}{
-		{
-			"upper left",
+			"top-left corner with zero height",
 			"testdata/pdf_handler_sample.pdf",
-			0, 0, 0, 0,
-			0, 792.0, 0, 0,
+			0, 0, 0,
+			0, 792.0, // For 612x792 page, top-left (0,0) with height 0 becomes (0, 792)
 		},
 		{
-			"bottom right",
+			"top-left corner with some height",
 			"testdata/pdf_handler_sample.pdf",
-			1, 1, 1, 1,
-			612.0, -792, 612.0, 792.0,
+			0, 0, 50,
+			0, 742.0, // For 612x792 page, top-left (0,0) with height 50 becomes (0, 792-0-50=742)
 		},
 		{
-			"Center of the page",
+			"center of page",
 			"testdata/pdf_handler_sample.pdf",
-			0.5, 0.5, 0.5, 0.5,
-			612.0 / 2, 0, 612.0 / 2, 792.0 / 2,
+			306, 396, 100,
+			306, 296, // For 612x792 page, center (306,396) with height 100 becomes (306, 792-396-100=296)
 		},
 		{
-			"upper left rotated",
-			"testdata/sample_rotate_90.pdf",
-			0, 0, 0, 0,
-			0, 612.0, 0, 0,
+			"bottom-right corner",
+			"testdata/pdf_handler_sample.pdf",
+			612, 792, 0,
+			612, 0, // For 612x792 page, bottom-right (612,792) with height 0 becomes (612, 792-792-0=0)
 		},
 		{
-			"bottom right rotated",
+			"rotated page top-left",
 			"testdata/sample_rotate_90.pdf",
-			1, 1, 1, 1,
-			792.0, -612, 792.0, 612.0,
+			0, 0, 0,
+			0, 612.0, // For 792x612 rotated page, top-left (0,0) with height 0 becomes (0, 612)
 		},
 		{
-			"Center of the rotated page",
+			"rotated page center",
 			"testdata/sample_rotate_90.pdf",
-			0.5, 0.5, 0.5, 0.5,
-			792.0 / 2, 0, 792.0 / 2, 612.0 / 2,
+			396, 306, 50,
+			396, 256, // For 792x612 rotated page, center (396,306) with height 50 becomes (396, 612-306-50=256)
 		},
 	}
 
@@ -232,25 +164,21 @@ func TestPdfHandler_TestLocationSizeToPdfPoints(t *testing.T) {
 			handler := setupPdfHandler(t)
 			handle := openTestPDF(t, tt.path)
 
-			X, Y, Width, Height, err := handler.LocationSizeToPdfPoints(
+			x, y, err := handler.ConvertTopLeftToBottomLeft(
 				context.Background(),
 				handle,
 				0,
-				tt.X,
-				tt.Y,
-				tt.Width,
-				tt.Height,
+				tt.x,
+				tt.y,
+				tt.height,
 			)
-			require.NoError(t, err, "Failed to convert percentages relative to page dimensions to PDF Point for file: %s", tt.path)
+			require.NoError(t, err, "Failed to convert top-left to bottom-left coordinates for file: %s", tt.path)
 
-			require.InDelta(t, tt.expectedX, X, 0.1, "Unexpected x for file: %s", tt.path)
-			require.InDelta(t, tt.expectedY, Y, 0.1, "Unexpected y for file: %s", tt.path)
-			require.InDelta(t, tt.expectedWidth, Width, 0.1, "Unexpected width for file: %s", tt.path)
-			require.InDelta(t, tt.expectedHeight, Height, 0.1, "Unexpected height for file: %s", tt.path)
+			require.InDelta(t, tt.expectedX, x, 0.1, "Unexpected x coordinate for file: %s", tt.path)
+			require.InDelta(t, tt.expectedY, y, 0.1, "Unexpected y coordinate for file: %s", tt.path)
 		})
 	}
 }
-
 func TestPdfHandler_GetPageSize_InvalidPage(t *testing.T) {
 	t.Parallel()
 
@@ -300,41 +228,7 @@ func TestPdfHandler_TestGetPageSize(t *testing.T) {
 			require.InDelta(t, tt.expectedHeight, size.Height, 0.1, "Unexpected height for file: %s", tt.path)
 		})
 	}
-}
 
-func setupPdfHandler(t *testing.T) PdfHandler {
-	t.Helper()
-
-	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-	return PdfHandler{Logger: logger}
-}
-
-func openTestPDF(t *testing.T, filePath string) *PdfDocument {
-	t.Helper()
-
-	handler := setupPdfHandler(t)
-	file, err := os.Open(filePath)
-	require.NoError(t, err)
-
-	document, err := handler.OpenPDF(file)
-	require.NoError(t, err)
-
-	t.Cleanup(func() {
-		require.NoError(t, file.Close())
-		require.NoError(t, handler.ClosePDF(document))
-	})
-
-	return document
-}
-
-func addImageAndSave(t *testing.T, handler PdfHandler, document *PdfDocument, params ImageParams, outputPath string) {
-	t.Helper()
-
-	err := handler.AddImageToPage(document, params)
-	require.NoError(t, err, "failed to add image")
-
-	err = handler.SavePDF(document, outputPath)
-	require.NoError(t, err, "failed to save PDF")
 }
 
 func TestPdfHandler_AddImageToPage(t *testing.T) {
@@ -352,14 +246,14 @@ func TestPdfHandler_AddImageToPage(t *testing.T) {
 			outputPath: "tmp/output_rotate_0_add_image_to_page_valid_image.pdf",
 			imageParams: ImageParams{
 				Page: 0,
-				Location: struct {
-					X float64
-					Y float64
-				}{X: 0, Y: 1 - 0.1452},
-				Size: struct {
-					Width  float64
-					Height float64
-				}{Width: 0.7239, Height: 0.1452},
+				Location: Location{
+					X: 0,
+					Y: 677,
+				},
+				Size: Size{
+					Width:  443,
+					Height: 115,
+				},
 				ImagePath: "testdata/test_signature.png",
 			},
 		},
@@ -369,14 +263,14 @@ func TestPdfHandler_AddImageToPage(t *testing.T) {
 			outputPath: "tmp/output_rotate_90_add_image_to_page_valid_image.pdf",
 			imageParams: ImageParams{
 				Page: 0,
-				Location: struct {
-					X float64
-					Y float64
-				}{X: 0, Y: 1 - 0.1452},
-				Size: struct {
-					Width  float64
-					Height float64
-				}{Width: 0.7239, Height: 0.1452},
+				Location: Location{
+					X: 0,
+					Y: 523,
+				},
+				Size: Size{
+					Width:  573,
+					Height: 88,
+				},
 				ImagePath: "testdata/test_signature.png",
 			},
 		},
@@ -386,14 +280,14 @@ func TestPdfHandler_AddImageToPage(t *testing.T) {
 			outputPath: "tmp/output_rotate_180_add_image_to_page_valid_image_top_right.pdf",
 			imageParams: ImageParams{
 				Page: 0,
-				Location: struct {
-					X float64
-					Y float64
-				}{X: 1 - 0.1810, Y: 0},
-				Size: struct {
-					Width  float64
-					Height float64
-				}{Width: 0.1810, Height: 0.0363},
+				Location: Location{
+					X: 501,
+					Y: 0,
+				},
+				Size: Size{
+					Width:  110,
+					Height: 28,
+				},
 				ImagePath: "testdata/test_signature.png",
 			},
 		},
@@ -403,14 +297,31 @@ func TestPdfHandler_AddImageToPage(t *testing.T) {
 			outputPath: "tmp/output_rotate_270_add_image_to_page_valid_image.pdf",
 			imageParams: ImageParams{
 				Page: 0,
-				Location: struct {
-					X float64
-					Y float64
-				}{X: 1 - 0.5593, Y: 1 - 0.1879},
-				Size: struct {
-					Width  float64
-					Height float64
-				}{Width: 0.5593, Height: 0.1879},
+				Location: Location{
+					X: 349,
+					Y: 497,
+				},
+				Size: Size{
+					Width:  443,
+					Height: 114,
+				},
+				ImagePath: "testdata/test_signature.png",
+			},
+		},
+		{
+			name:       "Valid Image - unreliable content stream",
+			pdfPath:    "testdata/sample_content.pdf",
+			outputPath: "tmp/output_add_image_to_unreliable_content_stream.pdf",
+			imageParams: ImageParams{
+				Page: 0,
+				Location: Location{
+					X: 0,
+					Y: 0,
+				},
+				Size: Size{
+					Width:  443,
+					Height: 115,
+				},
 				ImagePath: "testdata/test_signature.png",
 			},
 		},
@@ -446,14 +357,14 @@ func TestPdfHandler_AddImageToPage_InvalidPage(t *testing.T) {
 
 	params := ImageParams{
 		Page: 13,
-		Location: struct {
-			X float64
-			Y float64
-		}{X: 0, Y: 0},
-		Size: struct {
-			Width  float64
-			Height float64
-		}{Width: 0.1, Height: 0.1},
+		Location: Location{
+			X: 0,
+			Y: 0,
+		},
+		Size: Size{
+			Width:  61,
+			Height: 79,
+		},
 		ImagePath: "testdata/test_signature.png",
 	}
 
@@ -480,339 +391,21 @@ func TestPdfHandler_AddImageToPage_InvalidImage(t *testing.T) {
 
 	params := ImageParams{
 		Page: 0,
-		Location: struct {
-			X float64
-			Y float64
-		}{X: 0.5, Y: 0.1},
-		Size: struct {
-			Width  float64
-			Height float64
-		}{Width: 0.15, Height: 0.2},
+		Location: Location{
+			X: 306,
+			Y: 79,
+		},
+		Size: Size{
+			Width:  91,
+			Height: 158,
+		},
 		ImagePath: "testdata/test_signature-invalid.png",
 	}
 
 	err = handler.AddImageToPage(document, params)
 	require.Error(t, err)
 	require.Equal(t, "failure at the C/MuPDF add_image_to_page function: unknown image file format", err.Error())
-}
 
-func TestPdfHandler_TestGetFontAttributes_FontPath(t *testing.T) {
-	t.Parallel()
-
-	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-	handler := PdfHandler{Logger: logger}
-
-	tests := []struct {
-		name, fontName string
-		expectErr      bool
-		isStandardFont bool
-	}{
-		{"Standard Font Courier", "Courier", false, true},
-		{"Standard Font Courier-BoldOblique", "Courier-BoldOblique", false, true},
-		{"Standard Font ZapfDingbats", "ZapfDingbats", false, true},
-		{"Invalid Font", "NonExistentFont", true, false},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			fontPath, _, err := handler.getFontAttributes(context.Background(), tt.fontName, 0)
-			if tt.expectErr {
-				require.Error(t, err)
-			} else {
-				require.NoError(t, err)
-
-				if tt.isStandardFont {
-					require.Empty(t, fontPath, "Expected empty path for standard font %q", tt.fontName)
-				} else {
-					require.NotEmpty(t, fontPath, "Font path should not be empty for %q", tt.fontName)
-					if _, pathErr := os.Stat(fontPath); os.IsNotExist(pathErr) {
-						t.Errorf("Font path does not exist: %s", fontPath)
-					} else if pathErr != nil {
-						t.Errorf("Error checking font path: %v", pathErr)
-					} else {
-						t.Logf("Font path for %q: %s", tt.fontName, fontPath)
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestPdfHandler_TestGetFontAttributes_Descender(t *testing.T) {
-	t.Parallel()
-
-	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-	handler := PdfHandler{Logger: logger}
-
-	const epsilon = 0.05
-
-	tests := []struct {
-		name     string
-		fontName string
-		fontSize float64
-		expected float64
-	}{
-		{
-			name:     "Arial 12pt",
-			fontName: "Arial",
-			fontSize: 12.0,
-			expected: 2.547,
-		},
-		{
-			name:     "Times New Roman 10pt",
-			fontName: "Times New Roman",
-			fontSize: 10.0,
-			expected: 2.19,
-		},
-		{
-			name:     "Times New Roman 16pt",
-			fontName: "Times New Roman",
-			fontSize: 16.0,
-			expected: 3.469,
-		},
-		{
-			name:     "Courier 12pt",
-			fontName: "Courier",
-			fontSize: 12.0,
-			expected: 2.328,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			_, descender, err := handler.getFontAttributes(context.Background(), tt.fontName, tt.fontSize)
-			require.NoError(t, err)
-
-			if math.Abs(descender-tt.expected) > epsilon {
-				t.Errorf("got %.3f, expected %.3f ± %.2f", descender, tt.expected, epsilon)
-			}
-		})
-	}
-}
-
-func TestPdfHandler_AddTextBoxToPage(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name       string
-		inputFile  string
-		outputFile string
-		params     TextParams
-	}{
-		{
-			name:       "Text - A4 - Portrait - Times New Roman - 12",
-			inputFile:  "testdata/pdf_handler_sample.pdf",
-			outputFile: "tmp/output_rotate_0_add_text_to_page.pdf",
-			params: TextParams{
-				Value: "Hello, World!",
-				Page:  0,
-				Location: struct {
-					X float64
-					Y float64
-				}{X: 0, Y: 0.984},
-				Size: struct {
-					Width  float64
-					Height float64
-				}{Width: 0.7239, Height: 0.015},
-				Font: struct {
-					Family string
-					Size   float64
-				}{Family: "Times New Roman", Size: 12},
-			},
-		},
-		{
-			name:       "Text - A4 - Landscape - Times New Roman Italic - 8",
-			inputFile:  "testdata/sample_rotate_90.pdf",
-			outputFile: "tmp/output_rotate_90_add_text_to_page.pdf",
-			params: TextParams{
-				Value: "Hello, World!",
-				Page:  0,
-				Location: struct {
-					X float64
-					Y float64
-				}{X: 0, Y: 0.9866},
-				Size: struct {
-					Width  float64
-					Height float64
-				}{Width: 0.0617, Height: 0.0134},
-				Font: struct {
-					Family string
-					Size   float64
-				}{Family: "Times New Roman Italic", Size: 8},
-			},
-		},
-		{
-			name:       "Text - A4 - Landscape - Times New Roman Bold - 8 - top right",
-			inputFile:  "testdata/sample_rotate_270.pdf",
-			outputFile: "tmp/output_rotate_270_add_text_to_page_top_right.pdf",
-			params: TextParams{
-				Value: "Hello, World!",
-				Page:  0,
-				Location: struct {
-					X float64
-					Y float64
-				}{X: 1 - 0.063, Y: 0},
-				Size: struct {
-					Width  float64
-					Height float64
-				}{Width: 0.0617, Height: 0.0134},
-				Font: struct {
-					Family string
-					Size   float64
-				}{Family: "Times New Roman Bold", Size: 8},
-			},
-		},
-		{
-			name:       "Text - A4 - Portrait - Times New Roman - 24 - top right",
-			inputFile:  "testdata/sample_rotate_180.pdf",
-			outputFile: "tmp/output_rotate_180_add_text_to_page_top_right_24_fontsize.pdf",
-			params: TextParams{
-				Value: "Hello, World!",
-				Page:  0,
-				Location: struct {
-					X float64
-					Y float64
-				}{X: 1 - 0.294, Y: 0},
-				Size: struct {
-					Width  float64
-					Height float64
-				}{Width: 0.262, Height: 0.0285},
-				Font: struct {
-					Family string
-					Size   float64
-				}{Family: "Times New Roman", Size: 24},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-			handler := PdfHandler{Logger: logger}
-
-			file, err := os.Open(tt.inputFile)
-			require.NoError(t, err)
-			defer func() { require.NoError(t, file.Close()) }()
-
-			document, err := handler.OpenPDF(file)
-			require.NoError(t, err, "OpenPDF failed")
-			defer func() { require.NoError(t, handler.ClosePDF(document)) }()
-
-			err = handler.AddTextBoxToPage(document, tt.params)
-			require.NoError(t, err, "failed to add text")
-
-			err = handler.SavePDF(document, tt.outputFile)
-			require.NoError(t, err, "failed to save PDF")
-		})
-	}
-}
-
-func TestPdfHandler_AddTextBoxToPage_InvalidPage(t *testing.T) {
-	t.Parallel()
-
-	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-	handler := PdfHandler{Logger: logger}
-
-	file, err := os.Open("testdata/pdf_handler_sample.pdf")
-	require.NoError(t, err)
-	defer func() { require.NoError(t, file.Close()) }()
-
-	document, err := handler.OpenPDF(file)
-	if err != nil {
-		t.Fatalf("OpenPDF: %v", err)
-	}
-	defer func() { require.NoError(t, handler.ClosePDF(document)) }()
-
-	params := TextParams{
-		Value: "Hello, World!",
-		Page:  1,
-		Location: struct {
-			X float64
-			Y float64
-		}{X: 0, Y: 1},
-		Font: struct {
-			Family string
-			Size   float64
-		}{Family: "Courier", Size: 12},
-	}
-
-	err = handler.AddTextBoxToPage(document, params)
-	require.Error(t, err)
-	require.Equal(t, "failure at the AddTextBoxToPage function: failed to get page size: failure at the C/MuPDF get_page_size function: invalid page number: 2", err.Error())
-}
-
-func TestPdfHandler_AddTextBoxToPage_InvalidTextLengh(t *testing.T) {
-	t.Parallel()
-
-	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-	handler := PdfHandler{Logger: logger}
-
-	file, err := os.Open("testdata/pdf_handler_sample.pdf")
-	require.NoError(t, err)
-	defer func() { require.NoError(t, file.Close()) }()
-
-	document, err := handler.OpenPDF(file)
-	if err != nil {
-		t.Fatalf("OpenPDF: %v", err)
-	}
-	defer func() { require.NoError(t, handler.ClosePDF(document)) }()
-
-	params := TextParams{
-		Value: strings.Repeat("a", 301),
-		Page:  0,
-		Location: struct {
-			X float64
-			Y float64
-		}{X: 0, Y: 1},
-		Font: struct {
-			Family string
-			Size   float64
-		}{Family: "Courier", Size: 12},
-	}
-
-	err = handler.AddTextBoxToPage(document, params)
-	require.Error(t, err)
-	require.Equal(t, "failure at the C/MuPDF add_text_to_page function: Text exceeds maximum allowed size. Expected: 300, Actual: 301", err.Error())
-}
-
-func TestPdfHandler_AddTextBoxToPage_InvalidFont(t *testing.T) {
-	t.Parallel()
-
-	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-	handler := PdfHandler{Logger: logger}
-
-	file, err := os.Open("testdata/pdf_handler_sample.pdf")
-	require.NoError(t, err)
-	defer func() { require.NoError(t, file.Close()) }()
-
-	document, err := handler.OpenPDF(file)
-	if err != nil {
-		t.Fatalf("OpenPDF: %v", err)
-	}
-	defer func() { require.NoError(t, handler.ClosePDF(document)) }()
-
-	params := TextParams{
-		Value: "Hello, World!",
-		Page:  1,
-		Location: struct {
-			X float64
-			Y float64
-		}{X: 0, Y: 0},
-		Font: struct {
-			Family string
-			Size   float64
-		}{Family: "[not existing font]", Size: 12},
-	}
-
-	err = handler.AddTextBoxToPage(document, params)
-	require.Error(t, err)
-	require.Equal(t, "failure at PdfHandler AddTextBoxToPage function: failed to find font path for \"[not existing font]\"", err.Error())
 }
 
 func TestPdfHandler_AddCheckboxToPage(t *testing.T) {
@@ -831,14 +424,14 @@ func TestPdfHandler_AddCheckboxToPage(t *testing.T) {
 			params: CheckboxParams{
 				Value: true,
 				Page:  0,
-				Location: struct {
-					X float64
-					Y float64
-				}{X: 0, Y: 1 - 0.0253},
-				Size: struct {
-					Width  float64
-					Height float64
-				}{Width: 0.0327, Height: 0.0253},
+				Location: Location{
+					X: 0,
+					Y: 772,
+				},
+				Size: Size{
+					Width:  20,
+					Height: 20,
+				},
 			},
 		},
 		{
@@ -848,14 +441,14 @@ func TestPdfHandler_AddCheckboxToPage(t *testing.T) {
 			params: CheckboxParams{
 				Value: false,
 				Page:  0,
-				Location: struct {
-					X float64
-					Y float64
-				}{X: 0, Y: 1 - 0.0490},
-				Size: struct {
-					Width  float64
-					Height float64
-				}{Width: 0.0379, Height: 0.0490},
+				Location: Location{
+					X: 0,
+					Y: 582,
+				},
+				Size: Size{
+					Width:  30,
+					Height: 30,
+				},
 			},
 		},
 		{
@@ -865,14 +458,14 @@ func TestPdfHandler_AddCheckboxToPage(t *testing.T) {
 			params: CheckboxParams{
 				Value: true,
 				Page:  0,
-				Location: struct {
-					X float64
-					Y float64
-				}{X: 1 - 0.065, Y: 1 - 0.051},
-				Size: struct {
-					Width  float64
-					Height float64
-				}{Width: 0.065, Height: 0.051},
+				Location: Location{
+					X: 572,
+					Y: 751,
+				},
+				Size: Size{
+					Width:  40,
+					Height: 40,
+				},
 			},
 		},
 		{
@@ -882,14 +475,31 @@ func TestPdfHandler_AddCheckboxToPage(t *testing.T) {
 			params: CheckboxParams{
 				Value: true,
 				Page:  0,
-				Location: struct {
-					X float64
-					Y float64
-				}{X: 1 - 0.063, Y: 0},
-				Size: struct {
-					Width  float64
-					Height float64
-				}{Width: 0.063, Height: 0.082},
+				Location: Location{
+					X: 742,
+					Y: 0,
+				},
+				Size: Size{
+					Width:  50,
+					Height: 50,
+				},
+			},
+		},
+		{
+			name:       "Checkbox - unreliable content stream",
+			inputFile:  "testdata/sample_content.pdf",
+			outputFile: "tmp/output_add_checkbox_to_unreliable_content_stream.pdf",
+			params: CheckboxParams{
+				Value: true,
+				Page:  0,
+				Location: Location{
+					X: 573,
+					Y: 0,
+				},
+				Size: Size{
+					Width:  40,
+					Height: 40,
+				},
 			},
 		},
 	}
@@ -937,14 +547,14 @@ func TestPdfHandler_AddCheckboxToPage_InvalidPage(t *testing.T) {
 	params := CheckboxParams{
 		Value: true,
 		Page:  3,
-		Location: struct {
-			X float64
-			Y float64
-		}{X: 50, Y: 100},
-		Size: struct {
-			Width  float64
-			Height float64
-		}{Width: 20, Height: 20},
+		Location: Location{
+			X: 50,
+			Y: 100,
+		},
+		Size: Size{
+			Width:  20,
+			Height: 20,
+		},
 	}
 
 	err = handler.AddCheckboxToPage(document, params)
@@ -976,499 +586,62 @@ func TestPdfHandler_SavePDF_Valid(t *testing.T) {
 	}
 }
 
-func TestPdfHandler_MultipleOperations(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name       string
-		inputFile  string
-		outputFile string
-		operations []func(handler PdfHandler, document *PdfDocument) error
-	}{
-		{
-			name:       "Multiple operations on sample.pdf",
-			inputFile:  "testdata/pdf_handler_sample.pdf",
-			outputFile: "tmp/output_multiple_operations.pdf",
-			operations: []func(handler PdfHandler, document *PdfDocument) error{
-				func(handler PdfHandler, document *PdfDocument) error {
-					params := TextParams{
-						Value: "The quick brown fox jumps over the lazy dog!",
-						Page:  0,
-						Location: struct {
-							X float64
-							Y float64
-						}{X: 0.0817, Y: 0.0530},
-						Font: struct {
-							Family string
-							Size   float64
-						}{Family: "Courier", Size: 12},
-					}
-					return handler.AddTextBoxToPage(document, params)
-				},
-				func(handler PdfHandler, document *PdfDocument) error {
-					params := TextParams{
-						Value: "The quick brown fox jumps over the lazy dog!",
-						Page:  0,
-						Location: struct {
-							X float64
-							Y float64
-						}{X: 0.082, Y: 0.091},
-						Font: struct {
-							Family string
-							Size   float64
-						}{Family: "Courier", Size: 14},
-					}
-					return handler.AddTextBoxToPage(document, params)
-				},
-				func(handler PdfHandler, document *PdfDocument) error {
-					params := ImageParams{
-						Page: 0,
-						Location: struct {
-							X float64
-							Y float64
-						}{X: 0.163, Y: 0.179},
-						Size: struct {
-							Width  float64
-							Height float64
-						}{Width: 0.163, Height: 0.063},
-						ImagePath: "testdata/test_signature.png",
-					}
-					return handler.AddImageToPage(document, params)
-				},
-				func(handler PdfHandler, document *PdfDocument) error {
-					params := CheckboxParams{
-						Value: true,
-						Page:  0,
-						Location: struct {
-							X float64
-							Y float64
-						}{X: 0.245, Y: 0.242},
-						Size: struct {
-							Width  float64
-							Height float64
-						}{Width: 0.0327, Height: 0.0253},
-					}
-					return handler.AddCheckboxToPage(document, params)
-				},
-				func(handler PdfHandler, document *PdfDocument) error {
-					params := CheckboxParams{
-						Value: false,
-						Page:  0,
-						Location: struct {
-							X float64
-							Y float64
-						}{X: 0.245, Y: 0.280},
-						Size: struct {
-							Width  float64
-							Height float64
-						}{Width: 0.0327, Height: 0.0253},
-					}
-					return handler.AddCheckboxToPage(document, params)
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-			handler := PdfHandler{Logger: logger}
-
-			file, err := os.Open(tt.inputFile)
-			require.NoError(t, err)
-			defer func() { require.NoError(t, file.Close()) }()
-
-			document, err := handler.OpenPDF(file)
-			require.NoError(t, err, "OpenPDF failed")
-			defer func() { require.NoError(t, handler.ClosePDF(document)) }()
-
-			for _, operation := range tt.operations {
-				err := operation(handler, document)
-				require.NoError(t, err, "Operation failed")
-			}
-
-			err = handler.SavePDF(document, tt.outputFile)
-			require.NoError(t, err, "Failed to save PDF")
-		})
-	}
-}
-
-func TestPdfHandler_MultipleOperationsOnTextboxes(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name       string
-		inputFile  string
-		outputFile string
-		operations []func(handler PdfHandler, document *PdfDocument) error
-	}{
-		{
-			name:       "Multiple operations on texboxes.pdf",
-			inputFile:  "testdata/textboxes.pdf",
-			outputFile: "tmp/output_textboxes.pdf",
-			operations: []func(handler PdfHandler, document *PdfDocument) error{
-				func(handler PdfHandler, document *PdfDocument) error {
-					params := ImageParams{
-						Page: 0,
-						Location: struct {
-							X float64
-							Y float64
-						}{X: 0.0087145969, Y: 0.0033670034},
-						Size: struct {
-							Width  float64
-							Height float64
-						}{Width: 0.1633986928, Height: 0.0151515152},
-						ImagePath: "testdata/test_blue_box.png",
-					}
-					return handler.AddImageToPage(document, params)
-				},
-				func(handler PdfHandler, document *PdfDocument) error {
-					params := ImageParams{
-						Page: 0,
-						Location: struct {
-							X float64
-							Y float64
-						}{X: 0.0889615327, Y: 0.017957347},
-						Size: struct {
-							Width  float64
-							Height float64
-						}{Width: 0.1633986928, Height: 0.0151515152},
-						ImagePath: "testdata/test_blue_box.png",
-					}
-					return handler.AddImageToPage(document, params)
-				},
-				func(handler PdfHandler, document *PdfDocument) error {
-					params := ImageParams{
-						Page: 0,
-						Location: struct {
-							X float64
-							Y float64
-						}{X: 0.0087145969, Y: 0.962962963},
-						Size: struct {
-							Width  float64
-							Height float64
-						}{Width: 0.1633986928, Height: 0.0151515152},
-						ImagePath: "testdata/test_blue_box.png",
-					}
-					return handler.AddImageToPage(document, params)
-				},
-				func(handler PdfHandler, document *PdfDocument) error {
-					params := TextParams{
-						Value: "Qjstom",
-						Page:  0,
-						Location: struct {
-							X float64
-							Y float64
-						}{X: 0.0087145969, Y: 0.0033670034},
-						Size: struct {
-							Width  float64
-							Height float64
-						}{Width: 0.1633986928, Height: 0.0151515152},
-						Font: struct {
-							Family string
-							Size   float64
-						}{Family: "Times New Roman", Size: 12},
-					}
-					return handler.AddTextBoxToPage(document, params)
-				},
-				func(handler PdfHandler, document *PdfDocument) error {
-					params := TextParams{
-						Value: "qjWaAJj",
-						Page:  0,
-						Location: struct {
-							X float64
-							Y float64
-						}{X: 0.0889615327, Y: 0.017957347},
-						Size: struct {
-							Width  float64
-							Height float64
-						}{Width: 0.1633986928, Height: 0.0151515152},
-						Font: struct {
-							Family string
-							Size   float64
-						}{Family: "Times New Roman", Size: 12},
-					}
-					return handler.AddTextBoxToPage(document, params)
-				},
-				func(handler PdfHandler, document *PdfDocument) error {
-					params := TextParams{
-						Value: "QqWwJj",
-						Page:  0,
-						Location: struct {
-							X float64
-							Y float64
-						}{X: 0.0087145969, Y: 0.962962963},
-						Size: struct {
-							Width  float64
-							Height float64
-						}{Width: 0.1633986928, Height: 0.0151515152},
-						Font: struct {
-							Family string
-							Size   float64
-						}{Family: "Times New Roman", Size: 12},
-					}
-					return handler.AddTextBoxToPage(document, params)
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-			handler := PdfHandler{Logger: logger}
-
-			file, err := os.Open(tt.inputFile)
-			require.NoError(t, err)
-			defer func() { require.NoError(t, file.Close()) }()
-
-			document, err := handler.OpenPDF(file)
-			require.NoError(t, err, "OpenPDF failed")
-			defer func() { require.NoError(t, handler.ClosePDF(document)) }()
-
-			for _, operation := range tt.operations {
-				err := operation(handler, document)
-				require.NoError(t, err, "Operation failed")
-			}
-
-			err = handler.SavePDF(document, tt.outputFile)
-			require.NoError(t, err, "Failed to save PDF")
-		})
-	}
-}
-
-func TestPdfHandler_SaveToPNGOK(t *testing.T) {
-	t.Parallel()
-
-	for i := uint16(0); i < 13; i++ {
-		t.Run(fmt.Sprintf("page_%d", i), func(t *testing.T) {
-
-			logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-			handler := NewPdfHandler(context.Background(), logger)
-			file, err := os.Open("testdata/sample.pdf")
-			require.NoError(t, err)
-			defer func() { require.NoError(t, file.Close()) }()
-
-			document, err := handler.OpenPDF(file)
-			require.NoError(t, err)
-			defer func() { require.NoError(t, handler.ClosePDF(document)) }()
-
-			buf := bytes.NewBuffer([]byte{})
-			err = handler.SaveToPNG(document, i, 0, 0, 0, buf)
-			require.NoError(t, err)
-
-			resultPage, err := io.ReadAll(buf)
-			require.NoError(t, err)
-			expectedPage, err := os.ReadFile(fmt.Sprintf("testdata/sample_page%d.png", i))
-			require.NoError(t, err)
-			require.Equal(t, expectedPage, resultPage)
-		})
-	}
-}
-
-func BenchmarkPdfHandler_SaveToPNGPage0(b *testing.B)  { benchmarkPdfHandlerSaveToPNGRunner(0, b) }
-func BenchmarkPdfHandler_SaveToPNGPage1(b *testing.B)  { benchmarkPdfHandlerSaveToPNGRunner(1, b) }
-func BenchmarkPdfHandler_SaveToPNGPage2(b *testing.B)  { benchmarkPdfHandlerSaveToPNGRunner(2, b) }
-func BenchmarkPdfHandler_SaveToPNGPage3(b *testing.B)  { benchmarkPdfHandlerSaveToPNGRunner(3, b) }
-func BenchmarkPdfHandler_SaveToPNGPage4(b *testing.B)  { benchmarkPdfHandlerSaveToPNGRunner(4, b) }
-func BenchmarkPdfHandler_SaveToPNGPage5(b *testing.B)  { benchmarkPdfHandlerSaveToPNGRunner(5, b) }
-func BenchmarkPdfHandler_SaveToPNGPage6(b *testing.B)  { benchmarkPdfHandlerSaveToPNGRunner(6, b) }
-func BenchmarkPdfHandler_SaveToPNGPage7(b *testing.B)  { benchmarkPdfHandlerSaveToPNGRunner(7, b) }
-func BenchmarkPdfHandler_SaveToPNGPage8(b *testing.B)  { benchmarkPdfHandlerSaveToPNGRunner(8, b) }
-func BenchmarkPdfHandler_SaveToPNGPage9(b *testing.B)  { benchmarkPdfHandlerSaveToPNGRunner(9, b) }
-func BenchmarkPdfHandler_SaveToPNGPage10(b *testing.B) { benchmarkPdfHandlerSaveToPNGRunner(10, b) }
-func BenchmarkPdfHandler_SaveToPNGPage11(b *testing.B) { benchmarkPdfHandlerSaveToPNGRunner(11, b) }
-func BenchmarkPdfHandler_SaveToPNGPage12(b *testing.B) { benchmarkPdfHandlerSaveToPNGRunner(12, b) }
-
-func benchmarkPdfHandlerSaveToPNGRunner(page uint16, b *testing.B) {
-	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-	handler := NewPdfHandler(context.Background(), logger)
-
-	buf, err := os.ReadFile("testdata/sample.pdf")
-	require.NoError(b, err)
-
-	b.ReportAllocs()
-	for i := 0; i < b.N; i++ {
-		input := bytes.NewBuffer(buf)
-		document, err := handler.OpenPDF(input)
-		require.NoError(b, err)
-
-		output := bytes.NewBuffer([]byte{})
-		err = handler.SaveToPNG(document, page, 0, 0, 0, output)
-		require.NoError(b, err)
-
-		err = handler.ClosePDF(document)
-		require.NoError(b, err)
-	}
-}
-
-func TestPdfHandler_WrapPageContents(t *testing.T) {
-	t.Parallel()
+func setupPdfHandler(t *testing.T) PdfHandler {
+	t.Helper()
 
 	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-	handler := NewPdfHandler(context.Background(), logger)
+	return PdfHandler{Logger: logger}
+}
 
-	file, err := os.Open("testdata/pdf_handler_sample.pdf")
+func openTestPDF(t *testing.T, filePath string) *PdfDocument {
+	t.Helper()
+
+	handler := setupPdfHandler(t)
+	file, err := os.Open(filePath)
 	require.NoError(t, err)
-	defer func() { require.NoError(t, file.Close()) }()
 
 	document, err := handler.OpenPDF(file)
 	require.NoError(t, err)
-	defer func() { require.NoError(t, handler.ClosePDF(document)) }()
 
-	// Initially, no pages should be wrapped
-	require.False(t, document.wrappedPages[0])
+	t.Cleanup(func() {
+		require.NoError(t, file.Close())
+		require.NoError(t, handler.ClosePDF(document))
+	})
 
-	// First call to wrapPageContents for page 0 should mark it as wrapped
-	err = handler.wrapPageContents(context.Background(), document, 0)
-	require.NoError(t, err)
-	require.True(t, document.wrappedPages[0])
-
-	// Second call to wrapPageContents for page 0 should not error and page should still be marked as wrapped
-	err = handler.wrapPageContents(context.Background(), document, 0)
-	require.NoError(t, err)
-	require.True(t, document.wrappedPages[0])
-
-	// Test that other pages are not affected
-	require.False(t, document.wrappedPages[1])
-	require.False(t, document.wrappedPages[2])
+	return document
 }
 
-func TestPdfHandler_WrapPageContents_InvalidPage(t *testing.T) {
-	t.Parallel()
+func addImageAndSave(t *testing.T, handler PdfHandler, document *PdfDocument, params ImageParams, outputPath string) {
+	t.Helper()
 
-	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-	handler := NewPdfHandler(context.Background(), logger)
+	err := handler.AddImageToPage(document, params)
+	require.NoError(t, err, "failed to add image")
 
-	file, err := os.Open("testdata/pdf_handler_sample.pdf")
-	require.NoError(t, err)
-	defer func() { require.NoError(t, file.Close()) }()
-
-	document, err := handler.OpenPDF(file)
-	require.NoError(t, err)
-	defer func() { require.NoError(t, handler.ClosePDF(document)) }()
-
-	// Call wrapPageContents with invalid page number should return error
-	err = handler.wrapPageContents(context.Background(), document, 2)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "failure at wrap_page_contents_for_page")
+	err = handler.SavePDF(document, outputPath)
+	require.NoError(t, err, "failed to save PDF")
 }
 
-func TestPdfHandler_WrapPageContents_Integration(t *testing.T) {
-	t.Parallel()
-
+func BenchmarkPdfHandler_SoftSave(b *testing.B) {
 	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-	handler := NewPdfHandler(context.Background(), logger)
+	handler := PdfHandler{Logger: logger}
 
-	file, err := os.Open("testdata/sample.pdf")
-	require.NoError(t, err)
-	defer func() { require.NoError(t, file.Close()) }()
-
-	document, err := handler.OpenPDF(file)
-	require.NoError(t, err)
-	defer func() { require.NoError(t, handler.ClosePDF(document)) }()
-
-	// Test that annotation functions automatically wrap page contents
-	require.False(t, document.wrappedPages[0])
-	textParams := TextParams{
-		Value:    "Test Text",
-		Page:     0,
-		Location: Location{X: 0.1, Y: 0.1},
-		Size:     Size{Width: 0.3, Height: 0.05},
-		Font: struct {
-			Family string
-			Size   float64
-		}{
-			Family: "Times New Roman",
-			Size:   12,
-		},
-	}
-	err = handler.AddTextBoxToPage(document, textParams)
-	require.NoError(t, err)
-	require.True(t, document.wrappedPages[0])
-
-	// Test that annotation functions automatically wrap page contents
-	require.False(t, document.wrappedPages[1])
-	imageParams := ImageParams{
-		Page:      1,
-		Location:  Location{X: 0.5, Y: 0.5},
-		Size:      Size{Width: 0.2, Height: 0.2},
-		ImagePath: "testdata/test_signature.png",
-	}
-	err = handler.AddImageToPage(document, imageParams)
-	require.NoError(t, err)
-	require.True(t, document.wrappedPages[1]) // Should still be true
-
-	// Test that annotation functions automatically wrap page contents
-	require.False(t, document.wrappedPages[2])
-	checkboxParams := CheckboxParams{
-		Value:    true,
-		Page:     2,
-		Location: Location{X: 0.7, Y: 0.7},
-		Size:     Size{Width: 0.05, Height: 0.05},
-	}
-	err = handler.AddCheckboxToPage(document, checkboxParams)
-	require.NoError(t, err)
-	require.True(t, document.wrappedPages[2])
-}
-
-func BenchmarkPdfHandler_WrapPageContentsPerformance(b *testing.B) {
-	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-	handler := NewPdfHandler(context.Background(), logger)
-
-	file, err := os.Open("testdata/sample.pdf")
+	file, err := os.Open("testdata/00a730a587c1db1d59fa6dad689e2bf6a528cf98fd3a44891f193ef9bbc6a5cb.pdf")
 	require.NoError(b, err)
-	defer func() { require.NoError(b, file.Close()) }()
+	defer func() { _ = file.Close() }()
 
 	document, err := handler.OpenPDF(file)
 	require.NoError(b, err)
-	defer func() { require.NoError(b, handler.ClosePDF(document)) }()
+	defer func() { _ = handler.ClosePDF(document) }()
 
-	// Add first annotation to trigger initial wrap_page_contents call
-	// This is outside the benchmark timing
-	firstTextParams := TextParams{
-		Value:    "First annotation - setup",
-		Page:     0,
-		Location: Location{X: 0.1, Y: 0.1},
-		Size:     Size{Width: 0.3, Height: 0.05},
-		Font: struct {
-			Family string
-			Size   float64
-		}{
-			Family: "Times New Roman",
-			Size:   12,
-		},
-	}
-	err = handler.AddTextBoxToPage(document, firstTextParams)
-	require.NoError(b, err)
+	timestamp := time.Now().Format("20060102_150405")
+	tmpDir := filepath.Join("../../tmp", timestamp)
+	require.NoError(b, os.MkdirAll(tmpDir, 0o755))
 
-	require.True(b, document.wrappedPages[0], "Page should be wrapped after first annotation")
-
-	// Reset timer to exclude setup time
 	b.ResetTimer()
-	b.ReportAllocs()
-
-	// Benchmark loop - all subsequent annotations should be faster
-	// since wrapPageContents will return early (already wrapped)
 	for i := 0; i < b.N; i++ {
-		textParams := TextParams{
-			Value:    fmt.Sprintf("Benchmark annotation %d", i),
-			Page:     0,
-			Location: Location{X: 0.1, Y: 0.2 + float64(i%10)*0.05},
-			Size:     Size{Width: 0.3, Height: 0.04},
-			Font: struct {
-				Family string
-				Size   float64
-			}{
-				Family: "Times New Roman",
-				Size:   10,
-			},
-		}
-		err = handler.AddTextBoxToPage(document, textParams)
+		output := filepath.Join(tmpDir, fmt.Sprintf("save_%d.pdf", i))
+		err := handler.SavePDF(document, output)
 		require.NoError(b, err)
 	}
+	b.StopTimer()
 }
