+++ testdata/sample_content.pdf
@@ -1,0 +1,36 @@
+%PDF-1.7
+%µ¶
+
+1 0 obj
+<</Type/Catalog/Pages 2 0 R>>
+endobj
+
+2 0 obj
+<</Type/Pages/Kids[3 0 R]/Count 1>>
+endobj
+
+3 0 obj
+<</Type/Page/Parent 2 0 R/MediaBox[0 0 1500 1000]/Contents 4 0 R>>
+endobj
+
+4 0 obj
+<</Length 28>>
+stream
+0.24 0 0 -0.24 0 841.92 cm
+
+endstream
+endobj
+
+xref
+0 5
+0000000000 65535 f 
+0000000016 00000 n 
+0000000062 00000 n 
+0000000114 00000 n 
+0000000197 00000 n 
+
+trailer
+<</Size 5/Root 1 0 R>>
+startxref
+274
+%%EOF
